moved sp-metadata logic to npm
