+++ npm/src/typings.ts
@@ -391,6 +391,7 @@ export interface ISPSAMLConfig {
   }>;
   toMarkdown(): string;
   toHTML(): string;
+  toXMLMetadata(): Promise<string>;
 }
 
 export type DirectorySyncEventType =
